Refactor ReplyFeedback
